requirements.txt
python-telegram-bot>=20.0
google-generativeai
cryptography
requests
python-dotenv
aiosqlite
